updated user specific routes to check for session data

+++ flask_app/flask06.py
@@ -33,8 +33,10 @@ with app.app_context():
 @app.route('/')
 @app.route('/index')
 def index():
-    a_user = db.session.query(User).filter_by(email='jleiner1@uncc.edu').one()
-    return render_template('index.html', user = a_user)
+    if session.get('user'):
+    #a_user = db.session.query(User).filter_by(email='jleiner1@uncc.edu').one()
+        return render_template('index.html', user = session['user'])
+    return render_template('index.html')
 
 @app.route('/notes')
 def get_notes():
@@ -63,69 +65,79 @@ def get_note(note_id):
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
 
+    if session.get('user'):
     # check method used for request
-    if request.method == 'POST':
+        if request.method == 'POST':
 
         #get title data
-        title = request.form['title']
+            title = request.form['title']
 
         #get note data
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
 
         #update note data
-        note.title = title
-        note.text = text
+            note.title = title
+            note.text = text
         
         #update note in DB
-        db.session.add(note)
-        db.session.commit()
+            db.session.add(note)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:
+
+        #a_user = db.session.query(User).filter_by(email='jleiner1@uncc.edu').one()
+
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+
+        return render_template('new.html', note = my_note, user = session['user'])
     else:
-
-        a_user = db.session.query(User).filter_by(email='jleiner1@uncc.edu').one()
-
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note = my_note, user = a_user)
+        redirect(url_for('login'))
 
 @app.route('/notes/new', methods=['GET', 'POST'])
 def new_note():
 
-    if request.method == 'POST':
+    if session.get('user'):
+
+        if request.method == 'POST':
         
         # Get title data
-        title = request.form['title']
+            title = request.form['title']
 
         # Get note data
-        text = request.form['noteText']
+            text = request.form['noteText']
 
         # Create date stamp
-        from datetime import date
-        today = date.today()
+            from datetime import date
+            today = date.today()
 
         # Date format
-        today = today.strftime("%m-%d-%Y")
+            today = today.strftime("%m-%d-%Y")
 
-        new_record = Note(title, text, today)
-        db.session.add(new_record)
-        db.session.commit()
+            new_record = Note(title, text, today, session['user_id'])
+            db.session.add(new_record)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else:
+
+        #a_user = db.session.query(User).filter_by(email='jleiner1@uncc.edu').one()
+            return render_template('new.html', user = session['user'])
     else:
-
-        a_user = db.session.query(User).filter_by(email='jleiner1@uncc.edu').one()
-        return render_template('new.html', user = a_user)
+        return redirect(url_for('login'))
         
 @app.route('/notes/delete/<note_id>', methods=['POST'])
 def delete_note(note_id):
 
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
+        db.session.commit()
 
-    return redirect(url_for('get_notes'))
+        return redirect(url_for('get_notes'))
+    else:
+        return redirect(url_for('login'))
 
 @app.route('/register', methods=['POST', 'GET'])
 def register():
